Refactor cosigner package and interfaces
