+++ pkg/cosigner/icosigner.go
@@ -1,5 +1,1 @@
 package cosigner
-
-type iCosigner interface {
-	Health()
-}
